// src/types/flat.ts
export type Flat = {
  city: string;
  streetName: string;
  streetNumber: number;
  areaSize: number;
  hasAC: boolean;
  yearBuilt: number;
  rentPrice: number;
  dateAvailable: string;
  id: string;
  image?: File;
  imageUrl: string;
  userID: string;
  ownerFirstName: string;
  ownerLastName: string;
  ownerEmail: string;
};